requirements.txt
requests
teamcity-messages==1.7